chore: move location key helpers to project index
Relocated location-keys.js from src/shared/ into src/plugin/src/project-index/ so the helper lives with the project index and identifier-case features that consume it.

Updated plugin imports to match the new location.
