+++ src/plugin/src/project-index/location-keys.js
